+++ docs/reference/plugins/pydantic.rst
@@ -1,0 +1,5 @@
+pydantic
+========
+
+.. automodule:: litestar.plugins.pydantic
+    :members:
