+++ litestar/contrib/pydantic/__init__.py
@@ -1,20 +1,9 @@
+# ruff: noqa: TCH004, F401
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-from litestar.plugins import InitPluginProtocol
-
-from .pydantic_di_plugin import PydanticDIPlugin
-from .pydantic_dto_factory import PydanticDTO
-from .pydantic_init_plugin import PydanticInitPlugin
-from .pydantic_schema_plugin import PydanticSchemaPlugin
-
-if TYPE_CHECKING:
-    from pydantic import BaseModel
-    from pydantic.v1 import BaseModel as BaseModelV1
-
-    from litestar.config.app import AppConfig
-    from litestar.types.serialization import PydanticV1FieldsListType, PydanticV2FieldsListType
+from litestar.utils import warn_deprecation
 
 __all__ = (
     "PydanticDTO",
@@ -25,82 +14,35 @@ __all__ = (
 )
 
 
-def _model_dump(model: BaseModel | BaseModelV1, *, by_alias: bool = False) -> dict[str, Any]:
-    return (
-        model.model_dump(mode="json", by_alias=by_alias)  # pyright: ignore
-        if hasattr(model, "model_dump")
-        else {k: v.decode() if isinstance(v, bytes) else v for k, v in model.dict(by_alias=by_alias).items()}
-    )
-
-
-def _model_dump_json(model: BaseModel | BaseModelV1, by_alias: bool = False) -> str:
-    return (
-        model.model_dump_json(by_alias=by_alias)  # pyright: ignore
-        if hasattr(model, "model_dump_json")
-        else model.json(by_alias=by_alias)  # pyright: ignore
-    )
-
-
-class PydanticPlugin(InitPluginProtocol):
-    """A plugin that provides Pydantic integration."""
-
-    __slots__ = (
-        "exclude",
-        "exclude_defaults",
-        "exclude_none",
-        "exclude_unset",
-        "include",
-        "prefer_alias",
-        "validate_strict",
-    )
-
-    def __init__(
-        self,
-        exclude: PydanticV1FieldsListType | PydanticV2FieldsListType | None = None,
-        exclude_defaults: bool = False,
-        exclude_none: bool = False,
-        exclude_unset: bool = False,
-        include: PydanticV1FieldsListType | PydanticV2FieldsListType | None = None,
-        prefer_alias: bool = False,
-        validate_strict: bool = False,
-    ) -> None:
-        """Pydantic Plugin to support serialization / validation of Pydantic types / models
-
-        :param exclude: Fields to exclude during serialization
-        :param exclude_defaults: Fields to exclude during serialization when they are set to their default value
-        :param exclude_none: Fields to exclude during serialization when they are set to ``None``
-        :param exclude_unset: Fields to exclude during serialization when they arenot set
-        :param include: Fields to exclude during serialization
-        :param prefer_alias: Use the ``by_alias=True`` flag when dumping models
-        :param validate_strict: Use ``strict=True`` when calling ``.model_validate`` on Pydantic 2.x models
-        """
-        self.exclude = exclude
-        self.exclude_defaults = exclude_defaults
-        self.exclude_none = exclude_none
-        self.exclude_unset = exclude_unset
-        self.include = include
-        self.prefer_alias = prefer_alias
-        self.validate_strict = validate_strict
-
-    def on_app_init(self, app_config: AppConfig) -> AppConfig:
-        """Configure application for use with Pydantic.
-
-        Args:
-            app_config: The :class:`AppConfig <.config.app.AppConfig>` instance.
-        """
-        app_config.plugins.extend(
-            [
-                PydanticInitPlugin(
-                    exclude=self.exclude,
-                    exclude_defaults=self.exclude_defaults,
-                    exclude_none=self.exclude_none,
-                    exclude_unset=self.exclude_unset,
-                    include=self.include,
-                    prefer_alias=self.prefer_alias,
-                    validate_strict=self.validate_strict,
-                ),
-                PydanticSchemaPlugin(prefer_alias=self.prefer_alias),
-                PydanticDIPlugin(),
-            ]
+def __getattr__(attr_name: str) -> object:
+    if attr_name in __all__:
+        from litestar.plugins.pydantic import (
+            PydanticDIPlugin,
+            PydanticDTO,
+            PydanticInitPlugin,
+            PydanticPlugin,
+            PydanticSchemaPlugin,
         )
-        return app_config
+
+        warn_deprecation(
+            deprecated_name=f"litestar.contrib.pydantic.{attr_name}",
+            version="2.12",
+            kind="import",
+            removal_in="3.0",
+            info=f"importing {attr_name} from 'litestar.contrib.pydantic' is deprecated, please "
+            f"import it from 'litestar.plugins.pydantic' instead",
+        )
+        value = globals()[attr_name] = locals()[attr_name]
+        return value
+
+    raise AttributeError(f"module {__name__!r} has no attribute {attr_name!r}")  # pragma: no cover
+
+
+if TYPE_CHECKING:
+    from litestar.plugins.pydantic import (
+        PydanticDIPlugin,
+        PydanticDTO,
+        PydanticInitPlugin,
+        PydanticPlugin,
+        PydanticSchemaPlugin,
+    )
