+++ tests/unit/test_plugins/test_base.py
@@ -8,10 +8,10 @@ from click import Group
 from litestar import Litestar, MediaType, get
 from litestar.constants import UNDEFINED_SENTINELS
 from litestar.contrib.attrs import AttrsSchemaPlugin
-from litestar.contrib.pydantic import PydanticDIPlugin, PydanticInitPlugin, PydanticPlugin, PydanticSchemaPlugin
-from litestar.contrib.sqlalchemy.plugins import SQLAlchemySerializationPlugin
 from litestar.plugins import CLIPluginProtocol, InitPluginProtocol, OpenAPISchemaPlugin, PluginRegistry
 from litestar.plugins.core import MsgspecDIPlugin
+from litestar.plugins.pydantic import PydanticDIPlugin, PydanticInitPlugin, PydanticPlugin, PydanticSchemaPlugin
+from litestar.plugins.sqlalchemy import SQLAlchemySerializationPlugin
 from litestar.testing import create_test_client
 from litestar.typing import FieldDefinition
 
@@ -91,7 +91,7 @@ def test_plugin_registry_stringified_get() -> None:
     pydantic_plugin = PydanticPlugin()
     with pytest.raises(KeyError):
         PluginRegistry([CLIPlugin()]).get(
-            "litestar2.contrib.pydantic.PydanticPlugin"
+            "litestar2.plugins.pydantic.PydanticPlugin"
         )  # not a fqdn.  should fail # type: ignore[list-item]
         PluginRegistry([]).get("CLIPlugin")  # not a fqdn.  should fail # type: ignore[list-item]
 
@@ -99,7 +99,7 @@ def test_plugin_registry_stringified_get() -> None:
     assert PluginRegistry([cli_plugin, pydantic_plugin]).get(PydanticPlugin) is pydantic_plugin
     assert PluginRegistry([cli_plugin, pydantic_plugin]).get("PydanticPlugin") is pydantic_plugin
     assert (
-        PluginRegistry([cli_plugin, pydantic_plugin]).get("litestar.contrib.pydantic.PydanticPlugin") is pydantic_plugin
+        PluginRegistry([cli_plugin, pydantic_plugin]).get("litestar.plugins.pydantic.PydanticPlugin") is pydantic_plugin
     )
